Frontend Update
Added copy functionality for all chatbot pages

Fixed spacing in chatbot responses?

Updated user profile pic

a bunch of small UI changes

Removed a bunch of attributes to make consistent with the whole frontend. I may get back to it later, but it was a holdover from some tutorials I learned from, the attributes weren't really necessary for the scope of this project.

Consistency changes